rename common type dirs
