+++ packages/common/src/entities/ConsentCategory.ts
